Added Incomes pages and fixed deleting budgets (#41)
* Added entities

* Added migration

* Added controllers

* Updated state

* Added income service

* Added pages

* Added income to transactions pages

* Updated loading

* Added disable forms on loading

* Added remaining income

* Fixed delete budget

* Moved logout to end of the navbar

* Added delete confirmation

* Fixed tests

* PR tweak

+++ MyMoney.Web/ClientApp/src/app/pages/budgets/budgets.component.ts
@@ -61,6 +61,10 @@ export class BudgetsComponent implements OnInit {
    public get f() { return this.monthIdForm.controls; }
 
    public delete(id: number): void {
+      if (!confirm(`Delete budget ${id}?`)) {
+         return;
+      }
+
       this.budgetService.deleteBudget(id);
    }
